removed ugly logs

+++ app/src/main/java/com/pitstop/bluetooth/dataPackages/TripInfoPackage.java
@@ -11,7 +11,7 @@ public class TripInfoPackage {
     public String deviceId;
     public int tripId; // device specific
     public double mileage; // in kilometres
-    public double rtcTime; // unix time in seconds
+    public long rtcTime; // unix time in seconds
     public TripFlag flag;
 
     @Override
